+++ src/main/java/com/campestre/clube/backend_application/core/application/account/command/GetAccountByIdCommand.java
@@ -1,3 +1,3 @@
 package com.campestre.clube.backend_application.core.application.account.command;
 
-public record GetAccountByIdCommand(Integer id){}
+public record GetAccountByIdCommand(Long id){}
