+++ src/main/java/com/campestre/clube/backend_application/core/application/goal/command/GetGoalByTagIdCommand.java
@@ -1,3 +1,3 @@
 package com.campestre.clube.backend_application.core.application.goal.command;
 
-public record GetGoalByTagIdCommand(Integer tagId){}
+public record GetGoalByTagIdCommand(Long tagId){}
