+++ src/main/java/com/campestre/clube/backend_application/core/application/statement/command/GetStatementByIdCommand.java
@@ -1,3 +1,3 @@
 package com.campestre.clube.backend_application.core.application.statement.command;
 
-public record GetStatementByIdCommand(Integer id){}
+public record GetStatementByIdCommand(Long id){}
