+++ src/main/java/com/campestre/clube/backend_application/core/application/tag/command/DeleteTagCommand.java
@@ -1,3 +1,3 @@
 package com.campestre.clube.backend_application.core.application.tag.command;
 
-public record DeleteTagCommand(Integer id){}
+public record DeleteTagCommand(Long id){}
