+++ src/main/java/com/campestre/clube/backend_application/core/application/tag/command/GetTagByIdCommand.java
@@ -1,3 +1,3 @@
 package com.campestre.clube.backend_application.core.application.tag.command;
 
-public record GetTagByIdCommand(Integer id){}
+public record GetTagByIdCommand(Long id){}
