+++ src/main/java/com/campestre/clube/backend_application/core/application/transport/command/GetTransportByIdCommand.java
@@ -1,3 +1,3 @@
 package com.campestre.clube.backend_application.core.application.transport.command;
 
-public record GetTransportByIdCommand(Integer id){}
+public record GetTransportByIdCommand(Long id){}
