+++ src/main/java/com/campestre/clube/backend_application/core/domain/Account.java
@@ -4,13 +4,13 @@ import com.campestre.clube.backend_application.core.domain.enums.AccessTypeEnum;
 import com.campestre.clube.backend_application.core.domain.valueobject.Email;
 
 public class Account {
-    private Integer id;
+    private Long id;
     private Email email;
     private String password;
     private String name;
     private AccessTypeEnum access;
 
-    private Account(Integer id, Email email, String password, String name, AccessTypeEnum access) {
+    private Account(Long id, Email email, String password, String name, AccessTypeEnum access) {
         this.id = id;
         this.email = email;
         this.password = password;
@@ -18,7 +18,7 @@ public class Account {
         this.access = access;
     }
 
-    public static Account of(Integer id, String email, String password, String name, AccessTypeEnum access) {
+    public static Account of(Long id, String email, String password, String name, AccessTypeEnum access) {
         return new Account(
                 id,
                 Email.of(email),
@@ -38,7 +38,7 @@ public class Account {
         );
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
