+++ src/main/java/com/campestre/clube/backend_application/infrastructure/persistence/jpa/account/AccountJpaRepository.java
@@ -2,8 +2,8 @@ package com.campestre.clube.backend_application.infrastructure.persistence.jpa.a
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountJpaRepository extends JpaRepository<AccountEntity, Integer> {
-    Boolean existsByEmailAndIdNot(String email, Integer id);
+public interface AccountJpaRepository extends JpaRepository<AccountEntity, Long> {
+    Boolean existsByEmailAndIdNot(String email, Long id);
     Boolean existsByEmail(String email);
 
     AccountEntity findByEmail(String email);
