+++ src/main/java/com/campestre/clube/backend_application/infrastructure/persistence/jpa/tag/TagJpaRepository.java
@@ -2,11 +2,11 @@ package com.campestre.clube.backend_application.infrastructure.persistence.jpa.t
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TagJpaRepository extends JpaRepository<TagEntity, Integer> {
+public interface TagJpaRepository extends JpaRepository<TagEntity, Long> {
     boolean existsBySurnameIgnoreCase(String surname);
     boolean existsByColor(String color);
-    boolean existsBySurnameIgnoreCaseAndIdNot(String surname, Integer id);
-    boolean existsByColorAndIdNot(String color, Integer id);
+    boolean existsBySurnameIgnoreCaseAndIdNot(String surname, Long id);
+    boolean existsByColorAndIdNot(String color, Long id);
 
     TagEntity findBySurnameIgnoreCase(String surname);
 }
