+++ src/main/java/com/campestre/clube/backend_application/infrastructure/persistence/jpa/transport/TransportJpaRepository.java
@@ -1,16 +1,15 @@
 package com.campestre.clube.backend_application.infrastructure.persistence.jpa.transport;
 
-import com.campestre.clube.backend_application.core.domain.Transport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TransportJpaRepository extends JpaRepository<TransportEntity, Integer> {
+public interface TransportJpaRepository extends JpaRepository<TransportEntity, Long> {
     boolean existsByCompanyNameContainsIgnoreCaseAndCompanyNumberAndDriverNameContainsIgnoreCaseAndDriverNumber(
             String companyName, String companyNumber, String driverName, String driverNumber
     );
     boolean existsByCompanyNameContainsIgnoreCaseAndCompanyNumberAndDriverNameContainsIgnoreCaseAndDriverNumberAndIdNot(
-            String companyName, String companyNumber, String driverName, String driverNumber, Integer id
+            String companyName, String companyNumber, String driverName, String driverNumber, Long id
     );
     List<TransportEntity> findAllByOrderByRatingDesc();
 }
